nvidia: add Requires for max Xorg version
Lets use a Requires for the max Xorg version to ensure the Xorg ABI is
supported by the current driver release. If it's not, at least yum will
warn the user of an issue and not allow yum to update Xorg beyond
that supported by the driver and potentially break the system.

Signed-off-by: Philip J Perry <phil@elrepo.org>