alx: Fix multicast stream
I have a AR8161 and I tried using multicast with RHEL6.4 (through
ZeroConf).
The multicast stream was only readable by my app if the chipset was in
promiscuous mode.
So I searched the driver's source code and found a solution so that the
stream would not be filtered by the
driver (some code was disabled for RHEL6.4).

Contributed-by: A Roguez
Signed-off-by: Philip J Perry <phil@elrepo.org>