VirtualGL: build 32-bit libs package
This commit allows the 32-bit libs package to be built for RHEL7

Because there is no 32-bit tree on rhel7, the 32-bit libs package must
be built manually using 'rpmbuild -ba --target=i686 VirtualGL.spec' on
the command line. All 32-bit BRs will need to be manually installed on
the build host before building. Only VurtualGL-libs.i686 needs to be
distributed.

The 64-bit package(s) continues to build in mock as normal.

Signed-off-by: Philip J Perry <phil@elrepo.org>