Add kernel cmd line option 'plymouth.ignore-udev'
After installing the nvidia drivers, plymouth no longer functions, not even in text mode. Although Plymouth claims to fall back to text mode, it does not do so if it finds a DRM device.

It asks udev for a list of DRM or FB devices, and finds one provided by the 'nvidia' driver. However the nvidia DRM driver doesn't function as other DRM drivers do (i.e. it lacks real KMS or fbdev). At this point it should fall back to the text/console output but plymouth assumes that if it found a DRM device it should have worked, and so doesn't fall back to text/console. This often leads to a blank screen during boot, which is a problem for the systemd/PackageKit offline updates or if plymouth needs to ask for passwords.

This change sets the kernel command line option 'plymouth.ignore-udev' which causes plymouth to not ask dev for a list of DRM or FB devices at all. This means that plymouth starts in text mode and works correctly, even if it is just the fallback text mode. This is much better than having a blank screen during startup.

Existing solutions like setting the GRUB video mode or setting the vga=<mode> flags explicitly are not supported by the NVIDIA driver because they use a framebuffer (vesafb). Although this may appear to work, it can cause issues such as screen corruption. This change adds an option that lets both plymouth and NVIDIA work correctly.