yum-plugins-elrepo: reduce verbosity
Nobody actually wants to see elrepo's listing of every module. This
should be restricted to debug levels of verbosity. Otherwise, CentOS
users wind up with something amateurish looking, which isn't nice. It's
a shame this plugin must exist, but at least we can still make it stay
out of the way.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>