wireguard: add epoch to prevent installing on older kernels
An epoch that is equal to the minor point release has been added
to prevent accidentally installing a newer package version on an
older kernel.

As we try to simultaneously support the current RHEL release and
the older CentOS release at point release time, a newer package
update built against an older kernel could accidentally be installed.
Setting the epoch equal to the minor point release prevents this
from happening. The Epoch should be incremented at each point release
to match the point release kernel it is being built against.

Signed-off-by: Philip J Perry <phil@elrepo.org>